Factored out logic to deal with integer ranges
